backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
boto3==1.34.69
email-validator==2.1.1
python-dotenv==1.0.1
bcrypt
psycopg2-binary==2.9.9

